rename pyreinfolib to jiken
