+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "pyreinfolib"
+name = "jiken"
 version = "0.1.0"
-description = "Python library for accessing MLIT Real Estate Transaction Price Information API"
+description = "jiken - Where 地 (ji/land) meets 検証 (kensho/verification). Verify land values with official data from Japan's MLIT."
 readme = "README.md"
 requires-python = ">=3.11"
 license = { text = "MIT" }
@@ -31,16 +31,16 @@ dev = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/lfworks/pyreinfolib"
-Repository = "https://github.com/lfworks/pyreinfolib"
-Issues = "https://github.com/lfworks/pyreinfolib/issues"
+Homepage = "https://github.com/lfworks/jiken"
+Repository = "https://github.com/lfworks/jiken"
+Issues = "https://github.com/lfworks/jiken/issues"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/pyreinfolib"]
+packages = ["src/jiken"]
 
 # Ruff configuration
 [tool.ruff]
@@ -76,7 +76,7 @@ python_files = ["test_*.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
 addopts = [
-    "--cov=src/pyreinfolib",
+    "--cov=src/jiken",
     "--cov-report=term-missing",
     "--cov-report=html",
     "--cov-report=xml",
