+++ src/jiken/__init__.py
@@ -1,13 +1,13 @@
 __version__ = "0.1.0"
 
-from pyreinfolib.client import ReinfoLibClient
-from pyreinfolib.exceptions import (
+from jiken.client import ReinfoLibClient
+from jiken.exceptions import (
     ReinfoLibAPIError,
     ReinfoLibAuthError,
     ReinfoLibError,
     ReinfoLibRequestError,
 )
-from pyreinfolib.models import SearchCondition, Transaction
+from jiken.models import SearchCondition, Transaction
 
 __all__ = [
     "ReinfoLibClient",
